Created engine-test-utils package and moved engine updown there
