+++ src/engine/.gitignore
@@ -74,3 +74,6 @@ _deps
 test/helper_tests/map/outputs/*
 test/helper_tests/filter/outputs/*
 test/helper_tests/transformation/outputs/*
+
+# Vcpkg
+vcpkg/*
