+++ src/engine/test/engine-test-utils/pyproject.toml
@@ -3,10 +3,10 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "engine-updown"
+name = "engine-test-utils"
 version = "0.0.1"
 authors = [{ name = "Wazuh, Inc." }]
-description = "Test utility to launch and stop Engine instances"
+description = "Test utilities for Wazuh Engine"
 readme = "README.md"
 requires-python = ">=3.8"
 classifiers = ["Programming Language :: Python :: 3"]
